requirements.txt
dbt-snowflake
pandas
numpy
scikit-learn